requirements/runtime.txt
numpy
plotly
pandas
Pillow